% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhhs-package.R
\docType{package}
\name{dhhs-package}
\alias{dhhs}
\alias{dhhs-package}
\title{dhhs: Handle Common Tasks at the Department of Health and Human Services ('DHHS'), Victoria, Australia}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
