% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encode3202.R
\name{encode_3202}
\alias{encode_3202}
\alias{decode_3202}
\alias{Validate3202}
\title{Encode a character vector prefixed with 3202}
\usage{
encode_3202(x)

decode_3202(enc)

Validate3202(x)
}
\arguments{
\item{x, y}{\code{character(n)} Vectors to encode.}

\item{enc}{\code{integer(n)} Encoded vectors to decode.}
}
\value{
\code{encode_3202} returns an integer based on the digits following
the prefix \code{"3202"}. \code{decode_3202} is the reverse process.

\code{names2int} takes two character vectors of equal length and returns
an integer based on the first two digits. \code{Lookup4} is the inverse
process, returning the first two characters of each original vector
in \code{AaAa} case.
}
\description{
Encode a character vector prefixed with 3202
}
