% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeClassificationAcquired.R
\name{encode_ClassificationAcquired}
\alias{encode_ClassificationAcquired}
\alias{decode_ClassificationAcquired}
\alias{filter_Classification}
\title{Encode Classification and Acquired simultaneously}
\usage{
encode_ClassificationAcquired(Classification, Acquired)

decode_ClassificationAcquired(x)

filter_Classification(x, tbl)
}
\arguments{
\item{Classification, Acquired}{As in linelist.}

\item{x}{The encoded variable.}

\item{tbl}{A character vector of classifications}
}
\value{
\code{filter_Classification} returns a logical vector of columns where
the (decoded) Classification would be among the values in \code{tbl}.
}
\description{
Used to minimize memory use. Designed to accommodate filters
of the encoded variable
}
