% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_ClinicalStatus.R
\name{encode_ClinicalStatus}
\alias{encode_ClinicalStatus}
\title{Encode Clinical Status and Severity}
\usage{
encode_ClinicalStatus(x)
}
\arguments{
\item{x}{For encode, a character vector, the original \code{ClinicalStatus} or
\code{Severity}. For
decode, an raw vector, the encoded version.}
}
\value{
Clinical status and severity are encoded as a raw vector.
}
\description{
Encode Clinical Status and Severity
}
