% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeID.R
\name{encode_ID}
\alias{encode_ID}
\alias{fwalnume}
\alias{cipher_of}
\alias{decode_ID}
\title{Encode fixed width encoding}
\usage{
encode_ID(x, cipher = NULL, validate_cipher = TRUE, check_for_na = TRUE)

fwalnume(x, n = 18L)

cipher_of(e)

decode_ID(e, cipher = cipher_of(e))
}
\arguments{
\item{x}{A character vector, the id to be encoded, such as \code{"5002P00000DqpeuQAB"}.}

\item{cipher}{A list of two components.
The first, acharacter vector, whose length is the number of characters
in each identifier and each element of the vector comprises the number of
characters.

The second element of \code{cipher} is a boolean indicating whether or not
the original character vector had literal numbers (which will be preserved
if they are sufficiently small to be contained as \code{int32}).

If \code{NULL}, the default, the cipher is determined from \code{x}.}

\item{validate_cipher}{Should \code{cipher} be validated, ensuring that each element
of \code{x} is amenable to the \code{cipher} provided? Has no effect if
\code{cipher} is \code{NULL}. Setting this to \code{FALSE} with an incorrect
cipher is undefined behaviour.}

\item{check_for_na}{If \code{TRUE}, the default, \code{x} will be checked
for missing values first.}

\item{n}{A positive integer, the anticipated number of characters in each element of
\code{x}. Set to 18 by default since most identifiers are width 18.}

\item{e}{An encoded version of the ID.}
}
\value{
Each provides an interface to translate predictable character vectors
to integers and back.

\describe{
\item{\code{encode_ID}}{An integer vector, the encoded version of \code{x}. The attribute
\code{"dhhs_fwalnum_cipher"} of the object is the \strong{cipher}:
 a character vector whose length is the
number of characters in each element of the original and each element of which
gives the characters used at each position in every element of \code{x}.}
\item{\code{fwalnume}}{Returns the \code{cipher} determined by the contents of \code{x}.}
\item{\code{validate_fwalnume}}{If the cipher is valid, \code{0}; otherwise,
the position of the first invalid element.}
\item{\code{cipher_of}}{is a convenience function for extracting the cipher
from an encoded vector.}
}
}
\description{
Encodes fixed width identifiers like \code{AccountID} and
\code{RecordID} character vectors of alphanumeric strings.
}
