% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encode_linelist.R
\name{Encode_linelist}
\alias{Encode_linelist}
\alias{Decode_linelist}
\title{Encode/Decode using memory-light objects}
\usage{
Encode_linelist(DT, do_copy = TRUE)

Decode_linelist(DT, do_copy = TRUE)
}
\arguments{
\item{DT}{The \code{data.table} to encode/decode.}

\item{do_copy}{Whether \code{DT} should first be copied.}
}
\description{
Encode/Decode using memory-light objects
}
