% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_MetroRural.R
\name{encode_MetroRural}
\alias{encode_MetroRural}
\alias{decode_MetroRural}
\title{Encode MetroRural}
\usage{
encode_MetroRural(x)

decode_MetroRural(x)
}
\arguments{
\item{x}{For encode, a character vector; for decode, the encoded integer vector.}
}
\value{
Encode transforms
}
\description{
Encode MetroRural
}
