% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_PermitType.R
\name{encode_PermitType}
\alias{encode_PermitType}
\alias{decode_PermitType}
\title{Encode Permit Type}
\usage{
encode_PermitType(x)

decode_PermitType(x)
}
\arguments{
\item{x}{For encode, the character vector original permit type. For decode,
the decoded.}
}
\value{
\preformatted{
  NA = NA
   1 = Green
   2 = Orange
   3 = Red
}
}
\description{
Encode Permit Type
}
