% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_RecordType.R
\name{encode_RecordType}
\alias{encode_RecordType}
\alias{decode_RecordType}
\alias{get0_uRecordType}
\title{Decode Record type}
\usage{
encode_RecordType(x)

decode_RecordType(x)

get0_uRecordType()
}
\arguments{
\item{x}{For encode, a character vector, the original record type. For
decode, an integer vector, the encoded version.}
}
\value{
\code{get0_uRecordType()} returns the table of unique values expected.
}
\description{
Decode Record type
}
