% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeYN.R
\name{encode_YN}
\alias{encode_YN}
\alias{isEssentiallyLogical}
\title{Encode character vectors as logicals}
\usage{
encode_YN(x)

isEssentiallyLogical(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A logical vector with \code{TRUE} and \code{FALSE} matching Y/Yes and
N/No. Other values mapped to \code{NA}.

\code{isEssentiallyLogical} Returns \code{TRUE} or \code{FALSE}
based on whether the input numeric vector is comprised only of
0 and 1 (and NA).
}
\description{
Encode character vectors as logicals
}
