% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_nchar.R
\name{fast_nchar}
\alias{fast_nchar}
\title{Fast nchar that can be used on integers and strings}
\usage{
fast_nchar(x)
}
\arguments{
\item{x}{Either an integer or character vector.}
}
\value{
Same as \code{nchar(x)}
}
\description{
Fast nchar that can be used on integers and strings
}
