% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mystery-cases.R
\name{filter_mystery_cases}
\alias{filter_mystery_cases}
\title{Filter mystery cases}
\usage{
filter_mystery_cases(linelist)
}
\description{
A mystery case is one in which the classification is \code{"Confirmed"}
and which the \code{Acquired} field is either \code{"Acquired in Australia, unknown source"}
or \code{"Under investigation"}
}
