% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_STE.R
\name{filter_ste_in}
\alias{filter_ste_in}
\title{Filter by encoded state}
\usage{
filter_ste_in(x, y)
}
\arguments{
\item{x}{Either an atomic vector or a \code{data.table} with column \code{State}.}

\item{y}{The table to lookup.}
}
\value{
Either a logical vector or the filtered \code{data.table} the elements
of \code{x} in \code{y}.
}
\description{
Filter by encoded state
}
