% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepl_ClusterCategory.R
\name{grepl_ClusterCategory}
\alias{grepl_ClusterCategory}
\title{Find ClusterCategory pattern}
\usage{
grepl_ClusterCategory(x, table)
}
\arguments{
\item{x}{The character vector to search in.}

\item{table}{A character vector of terms to search for.}
}
\value{
Equivalent to \code{grepl(table, x, fixed = TRUE)}
or \code{grepl(paste0(table, collapse = "|"), x)} for sufficiently
fixed patterns in \code{table}.
}
\description{
Find a substring in a comma-separated string
}
\examples{
x <- c("Other", "Foo, Other", "Childcare, Education")
grepl_ClusterCategory(x, c("Education", "Other"))

}
