% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sitrep.R
\name{read_sitrep}
\alias{read_sitrep}
\title{Read COVID Sitrep file}
\usage{
read_sitrep(
  view,
  prev = FALSE,
  latest = FALSE,
  data_source = NULL,
  excl_diag_today = TRUE,
  use_fst = getOption("dhhs.use_fst", FALSE),
  columns = NULL,
  decode = TRUE,
  verbose = getOption("verbose", FALSE),
  reset_cache = getOption("dhhs.reset_cache", FALSE),
  exact = getOption("dhhs.sitrep_file_exact", FALSE)
)
}
\arguments{
\item{view, prev, latest}{As in `sitrep_file()` above.}

\item{data_source}{Which data to use?  One of \code{"today"}, \code{"previous"},
\code{"latest"}, or \code{"hour"}.

By default, the values of \code{previous,latest} are used. If \code{data_source}
is not \code{NULL} it takes priority.}

\item{excl_diag_today}{(TRUE or FALSE, default: \code{TRUE}) Should
rows where \code{DiagnosisDate >= today()} be dropped? Has no effect
if the requested file has no such column.}

\item{use_fst}{Should \code{.fst} files be used if available. If \code{TRUE},
but the corresponding file does not exist, it will be created and read in
on future reads.}

\item{columns}{Only useful when reading \code{.fst} files.
A character vector of column names to read in, dropping others. If \code{NULL},
the default, all columns are read.}

\item{decode}{If \code{FALSE}, \code{fst} files will be read without being decoded from
integer to character. Set to \code{FALSE} if you will decode the items
at a later stage and the performance penalty is too large.}

\item{verbose}{Be chatty and report the output from \code{fread(..., verbose = TRUE)}?}

\item{reset_cache}{If \code{TRUE}, saved \code{fst} files will be deleted. Useful
if the original files have been updated, or if the \code{fst} file had problems.}

\item{exact}{\code{FALSE | TRUE} If \code{TRUE}, then \code{view} must match exactly,
(up to file extensions).}
}
\value{
A \code{data.table} sourced from the requested sitrep file.
}
\description{
Used to read in sitrep files. Avoids
DRY violations throughout; hopefully with performance.
}
