% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sitrep.R
\name{sitrep_file}
\alias{sitrep_file}
\title{Sitrep file}
\usage{
sitrep_file(
  view = .views_avbl(),
  previous = FALSE,
  latest = FALSE,
  data_source = NULL,
  request_time = Sys.time(),
  fst = FALSE,
  must_exist = !fst,
  exact = FALSE
)
}
\arguments{
\item{view}{One of the (given) strings, which file to read. The argument is
called 'view' reflecting the SQL notion of a View.

\code{view} may also be a literal filename, in which case it is returned immediately
and other arguments will be ignored, silently.

Partial matches are accepted, and if multiple matches are plausible then the best
is chosen, with a message.}

\item{previous, latest}{(TRUE or FALSE, default: \code{FALSE}) Should the previous/latest
day's file be returned?}

\item{data_source}{Which data to use?  One of \code{"today"}, \code{"previous"},
\code{"latest"}, or \code{"hour"}.

By default, the values of \code{previous,latest} are used. If \code{data_source}
is not \code{NULL} it takes priority.}

\item{request_time}{(\code{POSIXct(1)}) The time requested, for the purposes
of identifying and invalidating existing caches.}

\item{fst}{(TRUE or FALSE, default: \code{FALSE}) Should the \code{.fst}
version be returned. If \code{FALSE}, the default, the path returned is the
standard \code{.txt} delimited/flat file.}

\item{must_exist}{(TRUE or FALSE, default: negation of \code{fst}) If \code{TRUE},
but the requested file does not exist, an error is raised. Otherwise, the putative
file location is returned.}

\item{exact}{\code{FALSE | TRUE} If \code{TRUE}, then \code{view} must match exactly,
(up to file extensions).}
}
\value{
A string, the path to the file requested.

A \code{data.table}, the file requested.
}
\description{
Convenience function to point to the correct location of
sitrep files on the E drive.
}
\details{
The following columns (if present) are modified in the following way:
\describe{
\item{\code{*Date}}{Any columns with 'Date' in their name will be coerced to \code{Date} class.}
\item{\code{Classification}}{Missing values are coalesced to \code{dhhs:::CLASSIFICIATION_MISSING_VALUE_ALIAS()}.}
\item{\code{AgeGroupTenYr}}{Missing values are coalesced to \code{dhhs:::AGEGROUP_MISSING_VALUE_ALIAS()}
  and "00-09" is mapped to "0-9".}
\item{\code{Latitude,Longitude}}{Any rows where \code{Latitude > 90} are assumed to be rows where the
latitude and longitude are confused/entered the wrong way and are switched back. Values of -999
are set to \code{NA_real_}.}
}
}
\author{
Hugh Parsonage
}
