% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_lga.R
\name{standardize_lga}
\alias{standardize_lga}
\alias{get_lga_names16}
\title{Standarize LGA Names}
\usage{
standardize_lga(x)

get_lga_names16(.dt = TRUE)
}
\arguments{
\item{x}{Character vector of names, probably LGA names.}

\item{.dt}{(Internal use only, whether to return a \code{data.table} instead of a character vector.)}
}
\value{
The character vector \code{x} but with best-guess standard names, where the LGA 
in \code{x} was not present in the LGA 2016 series.
}
\description{
Standarize LGA Names
}
