% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date2Int.R
\name{ddmmyyyy2Int}
\alias{ddmmyyyy2Int}
\title{Convert a date in dd/mm/YYYY form to a standard integer}
\usage{
ddmmyyyy2Int(x)
}
\arguments{
\item{x}{A date in \code{dd/mm/YYYY} format.}
}
\value{
An integer, equivalent to \code{as.integer(as.Date(x, format))}
for the format matching dd/mm/YYYY.
}
\description{
For efficient memory consumption
}
\note{
Missing values are mapped to \code{NA_integer_} as would be expected,
but so too are dates not in the above format, including for example
1/2/2000 (no leading zeroes).
}
