% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sitrep.R
\name{read_sitrep}
\alias{read_sitrep}
\title{Read COVID Sitrep file}
\usage{
read_sitrep(
  view,
  prev = FALSE,
  latest = FALSE,
  excl_diag_today = TRUE,
  use_fst = getOption("dhhs.use_fst", TRUE),
  columns = NULL,
  decode = TRUE,
  verbose = getOption("verbose", FALSE),
  reset_cache = getOption("dhhs.reset_cache", FALSE)
)
}
\arguments{
\item{view, prev, latest}{As in `sitrep_file()` above.}

\item{excl_diag_today}{(TRUE or FALSE, default: \code{TRUE}) Should
rows where \code{DiagnosisDate >= today()} be dropped? Has no effect
if the requested file has no such column.}

\item{use_fst}{Should \code{.fst} files be used if available. If \code{TRUE},
but the corresponding file does not exist, it will be created and read in
on future reads.}

\item{columns}{Only applicable when reading \code{.fst} files.
A character vector of column names to read in, dropping others. If \code{NULL},
the default, all columns are read.}

\item{decode}{If \code{FALSE}, \code{fst} files will be read without being decoded from
integer to character. Set to \code{FALSE} if you will decode the items
at a later stage and the performance penalty is too large.}

\item{verbose}{Be chatty and report the output from \code{fread(..., verbose = TRUE)}?}

\item{reset_cache}{If \code{TRUE}, saved \code{fst} files will be deleted. Useful
if the original files have been updated, or if the \code{fst} file had problems.}
}
\value{
A \code{data.table} sourced from the requested sitrep file.
}
\description{
Used to read in sitrep files. Avoids
DRY violations throughout; hopefully with performance.
}
